fix(render): 变更追踪精确到 (binding, array element) 粒度
同一 binding 的多 element write 各自独立追踪哈希，
避免第一个 element 每帧被误判为已变更而过量重写。

+++ GE/include/GE/Render/VulkanBase/VulkanDescriptorSet.h
@@ -103,8 +103,8 @@ class VulkanDescriptorSet
 
     std::vector<vk::WriteDescriptorSet>  m_WriteDescriptorSets;
 
-    /// 已写入的 bindings → 写入内容的哈希值。
-    std::unordered_map<uint32_t, size_t> m_UpdatedBindings;
+    /// 已写入的 (binding, array element) → 写入内容的哈希值。
+    std::unordered_map<uint64_t, size_t> m_UpdatedBindings;
 };
 
 } // namespace GE
